requirements.txt
streamlit
pandas
plotly
matplotlib
seaborn
logging
mysql-connector-python